+++ backend/src/main/java/org/spielhagen/backend/MemberRepository.java
@@ -9,11 +9,15 @@ import java.util.Optional;
 
 @Repository
 public interface MemberRepository extends MongoRepository<Member, String> {
-        Optional<Member> findTopByOrderByMitgliedsnummerDesc();
-        Optional<Member> findByMitgliedsnummer(Integer mitgliedsnummer);
-        // In Ihrem MemberRepository
-        List<Member> findByVornameContainingIgnoreCase(String vorname);
-        List<Member> findByNachnameContainingIgnoreCase(String nachname);
-        @Query("{ $text: { $search: ?0 } }")
-        List<Member> searchByTerms(String search);
+    Optional<Member> findTopByOrderByMitgliedsnummerDesc();
+
+    Optional<Member> findByMitgliedsnummer(Long mitgliedsnummer);
+
+    // In Ihrem MemberRepository
+    List<Member> findByVornameContainingIgnoreCase(String vorname);
+
+    List<Member> findByNachnameContainingIgnoreCase(String nachname);
+
+    @Query("{ $text: { $search: ?0 } }")
+    List<Member> searchByTerms(String search);
 }
